Name of map variable was changed. Type of quantity was changed. FileValidator became a Consumer. Method readFromFile was simplyfied.

+++ src/main/java/core/basesyntax/operations/OperationHandler.java
@@ -1,8 +1,7 @@
 package core.basesyntax.operations;
 
 import core.basesyntax.model.Fruit;
-import java.math.BigDecimal;
 
 public interface OperationHandler {
-    BigDecimal perform(Fruit fruit, BigDecimal operationQuantity);
+    long perform(Fruit fruit, long operationQuantity);
 }
